Fix GCS backend from https://github.com/mozilla/sccache/pull/212

+++ src/cache/gcs.rs
@@ -375,6 +375,6 @@ impl Storage for GCSCache {
         format!("GCS, bucket: {}", self.bucket)
     }
 
-    fn current_size(&self) -> Option<usize> { None }
-    fn max_size(&self) -> Option<usize> { None }
+    fn current_size(&self) -> Option<u64> { None }
+    fn max_size(&self) -> Option<u64> { None }
 }
